Fix two bugs in find words puzzle.
- Bug 1: sentence may be too long, there is no solution for the puzzle.
Expected: sentence should be limited to less than 18 characters. If it's more than that, regenerate.

- Bug 2: sentence may include characters that the student hasn't learned to read before
This is a tricky bug, balancing a few things

Given a set of characters to study (a small set of 5-10), and a set of characters the student learned (a large set of 100-300). Here are the constraints
- The sentence need to make sense for a primary school student
- The sentence should include the characters from the study set as much as possible
- The sentence can include characters from the learned set, but cannot include characters outside of both sets

Need to design an algorithm to make this work.

First implement a test case so I can test this out.